+++ backend/test/testLocalWork.js
@@ -1,0 +1,16 @@
+var FORMATTER = require('../../ethStorage/format/formatter.js');
+var mocha = require('mocha');
+var describe = mocha.describe;
+var it = mocha.it;
+var assert = require('chai').assert;
+var expect = require('chai').expect;
+
+// var FORMATTER = new StateDB(Settings.dbPath);
+
+
+describe('Test for local work setup',function () {
+    it('Geth blockchain connection',function () {
+        assert.equal(3,3)
+    });
+});
+
